move Jest tests outside src
