+++ test/advImageApi.test.js
@@ -1,22 +1,15 @@
 const {
-  searchApiUrlTest,
   fetchImages,
   fetchAlbum,
   fetchProcessImages,
   fetchTagImages
-} = require('./advImageApi');
+} = require('../src/utils/advImageApi');
 
-const param = '878753873422768:Hk5O_AutOONuI-2U4F05x5pVDTY';
-
-test('check search URL', () => {
-  const expectedValue = '878753873422768:Hk5O_AutOONuI-2U4F05x5pVDTY@api.cloudinary.com/v1_1/cantimaginewhy/resources/search'
-  expect(searchApiUrlTest(param)).toBe(expectedValue);
-})
 
 test('check new expression', () => {
   const expression = 'folder=icon';
   const expectedResults = 6;
-  return fetchImages(expression, 10, param).then(response => {
+  return fetchImages(expression, 10).then(response => {
     expect(response.resources.length).toBe(expectedResults);
   })
 })
@@ -24,7 +17,7 @@ test('check new expression', () => {
 test('contents of album', () => {
   const albumName = 'People';
   const expectedResults = 17;
-  return fetchAlbum(albumName, param).then(response => {
+  return fetchAlbum(albumName).then(response => {
     expect(response.resources.length).toBe(expectedResults)
   })
 })
@@ -32,14 +25,14 @@ test('contents of album', () => {
 test('process images', () => {
   const refKey = 'watertown_dam';
   const expectedResults = 5;
-  return fetchProcessImages(refKey, param).then(response => {
+  return fetchProcessImages(refKey).then(response => {
     expect(response.resources.length).toBe(expectedResults);
   })
 })
 
 test('tagged images', () => {
   const tag = 'en plein air';
-  return fetchTagImages(tag, param).then(response => {
+  return fetchTagImages(tag).then(response => {
     expect(response.resources.length).toBe(42);
     expect(response.resources[0].public_id).toBe('art/a_friday_in_september');
   })
